Refactor write checkpoints to make testing easier.
